Move cli tests in multiple files
